requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
celery-redbeat==0.10.0
Django==1.11.4
django-recurrence==1.4.1
djangorestframework==3.6.3
kombu==4.1.0
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
six==1.10.0
vine==1.1.4
